refactor: move app_only_queryPersonas
