+++ packages/mask/background/services/identity/index.ts
@@ -4,7 +4,7 @@ export {
     createPersonaByMnemonic,
     createPersonaByMnemonicV2,
 } from './persona/create'
-export { mobile_queryPersonaRecordsFromIndexedDB } from './persona/query'
+export { mobile_queryPersonaRecordsFromIndexedDB, mobile_queryPersonas } from './persona/query'
 export {
     deletePersona,
     logoutPersona,
